fix(service): added isNotEmpty guards

+++ src/main/kotlin/org/careerseekers/cseventsservice/services/reports/AllChildrenReportService.kt
@@ -82,7 +82,9 @@ class AllChildrenReportService(
         }
 
         val rows = deferredRows.awaitAll().also {
-            childToDirectionService.deleteAllByIds(removableRecords.map { it.id })
+            if (removableRecords.isNotEmpty()) {
+                childToDirectionService.deleteAllByIds(removableRecords.map { it.id })
+            }
         }
 
         return@coroutineScope createExcelFile(rows)
